Server

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,5 +1,7 @@
-# execute pkill command to kill bashscript file killmenow
+# kill proccess killmenow
+
 exec { 'pkill':
   command => 'pkill -9 -f killmenow',
-  path    => ['/usr/bin', '/usr/sbin', '/bin']
+  path    => '/usr/bin:/usr/sbin:/bin:/sbin',
+  onlyif  => 'pgrep -f killmenow',
 }
